Genre/Mood Browsing
Did this as a project to learn React. Scraped the genres and moods
from the dropdowns in the edit track window on Audius, so they
aren't updating from the API and will need to be manually updated
if the options on the site change. The artist name links to their
page, and the track name links to the track page. Clicking the
track image plays the track, clicking it again stops it. No play
progress bar or volume controls implemented since this was just
a demo playground for me.